04 Cloud/serverless/requirements.txt
functions-framework==3.*
requests==2.31.0
google-cloud-firestore==2.15.0 